Extract bucket setup operations out of r2.py into bucket_setup.py
ensure_bucket, create_inboxes, apply_cors, enable_r2_dev_public, and
set_public_policy are one-time setup operations — not used at runtime.
Moving them removes httpx, shutil, subprocess, and urlparse from r2.py.
